Refactor Broadcast -> Replication
