+++ nomos-da/network/core/src/lib.rs
@@ -1,4 +1,4 @@
-mod broadcast;
 mod dispersal;
 mod protocol;
+mod replication;
 mod sampling;
